SpreadsheetFormulaWidget.saveFormulaText refactoring
- Inlined previous SpreadsheetViewportWidget.saveFormulaText
